ref(onboarding): Split donet onboarding docs
