+++ static/app/gettingStartedDocs/dotnet/dotnet/index.tsx
@@ -1,0 +1,15 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+
+import {crashReport} from './crashReport';
+import {feedback} from './feedback';
+import {onboarding} from './onboarding';
+import {profiling} from './profiling';
+
+const docs: Docs = {
+  onboarding,
+  feedbackOnboardingCrashApi: feedback,
+  crashReportOnboarding: crashReport,
+  profilingOnboarding: profiling,
+};
+
+export default docs;
