+++ static/app/gettingStartedDocs/dotnet/dotnet/onboarding.spec.tsx
@@ -4,7 +4,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './dotnet';
+import docs from '.';
 
 describe('dotnet onboarding docs', () => {
   it('renders errors onboarding docs correctly', async () => {
