+++ static/app/gettingStartedDocs/dotnet/dotnet/profiling.tsx
@@ -1,0 +1,8 @@
+import {getDotnetProfilingOnboarding} from 'sentry/utils/gettingStartedDocs/dotnet';
+
+import {getInstallSnippetCoreCli, getInstallSnippetPackageManager} from './onboarding';
+
+export const profiling = getDotnetProfilingOnboarding({
+  getInstallSnippetPackageManager,
+  getInstallSnippetCoreCli,
+});
